If overshooting tops feature returns an empty string, strip trailing comma
In cases where the overshooting top feature is present but its lat/lon
positions are masked, plotOT() returns an empty string. If this is the
case then a trailing comma should be removed from the list of features
before it is returned.

Fix for trailing commma issue: https://github.com/cemac/ADAGUC-utilities/issues/2